continue refactor into different modules
