+++ src/lib.rs
@@ -1,1 +1,4 @@
+pub mod maus;
 pub mod sysfs;
+pub mod auto;
+pub mod dummy;
